+++ tests/testthat/test_MsBackendCompDb-functions.R
@@ -1,18 +1,6 @@
-test_that(".valid_spectra_data_required_columns works", {
-    df <- DataFrame()
-    expect_null(.valid_spectra_data_required_columns(df))
-    df <- DataFrame(msLevel = 1L)
-    expect_match(.valid_spectra_data_required_columns(df),
-                 "Required column")
-    df$dataStorage <- "some"
-    expect_null(.valid_spectra_data_required_columns(df))
-})
-
-test_that(".valid_ms_backend_dbcon works", {
-    res <- .valid_ms_backend_dbcon(cmp_db@dbcon)
-    expect_true(length(res) == 1)
-    res <- .valid_ms_backend_dbcon(cmp_spctra_db@dbcon)
-    expect_true(length(res) == 0)
+test_that(".valid_dbcon MsBackendCompDb works", {
+    expect_equal(CompoundDb:::.valid_dbcon(cdb@dbcon), NULL)
+    expect_match(CompoundDb:::.valid_dbcon(cmp_db@dbcon), "no MS/MS")
 })
 
 test_that("MsBackendCompDb works", {
@@ -21,25 +9,72 @@ test_that("MsBackendCompDb works", {
     expect_true(is.null(res@dbcon))
 })
 
-test_that(".peaks works", {
-    be <- Spectra(cmp_spctra_db)@backend
-    pks <- .peaks(be)
-    expect_equal(names(pks), as.character(be$spectrum_id))
-    expect_true(is.list(pks))
-    expect_true(is.matrix(pks[[1]]))
-    expect_true(all(colnames(pks[[2]]) == c("mz", "intensity")))
+test_that(".map_spectraVariables_to_sql works", {
+    res <- .map_spectraVariables_to_sql("msLevel")
+    expect_equal(res, "ms_level")
+    res <- .map_spectraVariables_to_sql(c("mz", "collisionEnergy", "ms_level"))
+    expect_equal(res, c("mz", "collision_energy", "ms_level"))
+})
 
-    be <- be[c(3, 4, 2)]
-    pks <- .peaks(be)
-    expect_equal(names(pks), as.character(be$spectrum_id))
+test_that(".map_sql_to_spectraVariables works", {
+    res <- .map_sql_to_spectraVariables("ms_level")
+    expect_equal(res, "msLevel")
+    res <- .map_sql_to_spectraVariables(c("precursor_mz", "intensity"))
+    expect_equal(res, c("precursorMz", "intensity"))
+})
 
-    mzs <- .peaks(be, column = "mz")
-    expect_true(is.list(mzs))
-    expect_true(is.numeric(mzs[[2]]))
-    expect_identical(mzs, lapply(pks, function(z) z[, 1]))
+test_that(".fetch_peaks works", {
+    res <- CompoundDb:::.fetch_peaks(MsBackendCompDb())
+    expect_true(is.data.frame(res))
+    expect_true(nrow(res) == 0)
+    expect_true(all(colnames(res) %in% c("spectrum_id", "mz", "intensity")))
 
-    ints <- .peaks(be, column = "intensity")
-    expect_true(is.list(ints))
-    expect_true(is.numeric(ints[[2]]))
-    expect_identical(ints, lapply(pks, function(z) z[, 2]))
+    be <- backendInitialize(MsBackendCompDb(), cdb)
+    res <- CompoundDb:::.fetch_peaks(be)
+    expect_true(is.data.frame(res))
+    expect_true(nrow(res) > 0)
+    expect_true(all(colnames(res) %in% c("spectrum_id", "mz", "intensity")))
+})
+
+test_that(".peaks_data works", {
+    res <- CompoundDb:::.peaks_data(MsBackendCompDb())
+    expect_true(length(res) == 0)
+    expect_true(is.list(res))
+
+    be <- backendInitialize(MsBackendCompDb(), cdb)
+    pks <- CompoundDb:::.fetch_peaks(be)
+    res <- CompoundDb:::.peaks_data(be)
+    expect_true(length(res) > 1)
+    expect_true(is.matrix(res[[1L]]))
+})
+
+test_that(".spectra_data works", {
+    res <- CompoundDb:::.spectra_data(MsBackendCompDb())
+    expect_true(nrow(res) == 0)
+    expect_true(is(res, "DataFrame"))
+    expect_equal(colnames(res), spectraVariables(MsBackendCompDb()))
+
+    be <- backendInitialize(MsBackendCompDb(), cdb)
+    res <- CompoundDb:::.spectra_data(be)
+    expect_true(nrow(res) == length(be))
+    expect_equal(colnames(res), spectraVariables(be))
+    expect_equal(be@spectraIds, as.character(res$spectrum_id))
+
+    expect_true(is(res$mz, "NumericList"))
+
+    res_2 <- CompoundDb:::.spectra_data(be, "mz")
+    expect_true(is(res_2, "DataFrame"))
+    expect_true(colnames(res_2) == "mz")
+    expect_equal(res$mz, res_2$mz)
+
+    res_2 <- CompoundDb:::.spectra_data(be, c("name", "compound_id"))
+    expect_true(is(res_2, "DataFrame"))
+    expect_true(all(colnames(res_2) == c("name", "compound_id")))
+    expect_equal(nrow(res_2), length(be))
+
+    be <- backendInitialize(MsBackendCompDb(), cmp_spctra_db)
+    res_2 <- CompoundDb:::.spectra_data(be, c("name", "compound_id"))
+    expect_true(is(res_2, "DataFrame"))
+    expect_true(all(colnames(res_2) == c("name", "compound_id")))
+    expect_equal(nrow(res_2), length(be))
 })